A global function is created to lock in each subscription and not repeat code
