+++ src/libs/auth/auth.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { HttpModule } from '@nestjs/axios';
 import { ConfigModule } from '@nestjs/config';
-import { AuthService } from './auth.service';
+import { ApiService } from './auth.service';
 
 @Module({
   imports: [HttpModule, ConfigModule],
-  providers: [AuthService],
-  exports: [AuthService],
+  providers: [ApiService],
+  exports: [ApiService],
 })
 export class AuthModule {}
