+++ src/module/iframes/iframes.service.ts
@@ -1,10 +1,5 @@
 import { HttpService } from '@nestjs/axios';
-import {
-  HttpException,
-  HttpStatus,
-  Injectable,
-  NotFoundException,
-} from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { CreateIframeDto } from './dto/create-iframe.dto';
@@ -12,6 +7,7 @@ import { UpdateIframeDto } from './dto/update-iframe.dto';
 import { IframeSuscription } from './entities/iframe.entity';
 import { User } from '../users/entities/user.entity';
 import { lastValueFrom } from 'rxjs';
+import { ApiService } from 'src/libs/auth/auth.service';
 
 @Injectable()
 export class IframesService {
@@ -21,43 +17,11 @@ export class IframesService {
     private readonly iframeModel: Model<IframeSuscription>,
     @InjectModel(User.name) private readonly userModel: Model<User>,
     private readonly httpService: HttpService,
+    private readonly apiService: ApiService,
   ) {}
 
   async create(createIframeDto: CreateIframeDto): Promise<void> {
-    const createApiKeyParams = {
-      type: 'iframe',
-      usageCount: 0,
-      limit: 100,
-    };
-
-    const headers = {
-      'x-api-key': 'p5ypxpbidn0200uvh4cz0plx3n2zqy',
-    };
-
-    const apiKeyResponse = await lastValueFrom(
-      this.httpService.post(
-        'http://localhost:4000/key-subscription/new',
-        createApiKeyParams,
-        { headers },
-      ),
-    );
-    console.log('hola', apiKeyResponse.data);
-
-    if (apiKeyResponse.status !== 201) {
-      console.log('hola');
-      throw new HttpException(
-        'Failed to create API key',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
-    } else if (apiKeyResponse.data.apiKey === undefined) {
-      console.log('hola2');
-      throw new HttpException(
-        'Failed to create API key',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
-    }
-    const apiKey = apiKeyResponse.data.apiKey;
-    console.log('API Key:', apiKey);
+    const apiKey = await this.apiService.getApiKey('iframe');
 
     createIframeDto.apikey = apiKey;
